Add unit tests, and refactor to support them.
